LMS web/pyproject.toml
[project]
name = "lms-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "config>=0.5.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
]
